Resolved phpstan, phpcs issues. Added new github actions (Magento 2.4.7, 2.4.8)

+++ Test/Unit/Ui/Component/Listing/CustomerPrice/Column/PriceTest.php
@@ -104,8 +104,8 @@ class PriceTest extends TestCase
             ->method('toCurrency')
             ->will(
                 $this->returnValueMap([
-                    [sprintf("%f", 10.00), [], '$10.00'],
-                    [sprintf("%f", 20.00), [], '$20.00']
+                    [(float)sprintf("%f", 10.00), [], '$10.00'],
+                    [(float)sprintf("%f", 20.00), [], '$20.00']
                 ])
             );
